svn:revision:68100
svn:branch:RELEASE
svn:account:starseeker

+++ src/conv/iges/summary.c
@@ -24,8 +24,8 @@
 void
 Summary()
 {
-    int i;
-    int indep_entities = 0;
+    size_t i;
+    size_t indep_entities = 0;
 
     bu_log("Summary of entity types found:\n");
     for (i = 0; i <= ntypes; i++) {
@@ -40,14 +40,14 @@ Summary()
 	if (!subord)
 	    indep_entities++;
     }
-    bu_log("%d Independent entities\n", indep_entities);
+    bu_log("%zu Independent entities\n", indep_entities);
 }
 
 
 void
 Zero_counts()
 {
-    int i;
+    size_t i;
 
     for (i = 0; i <= ntypes; i++)
 	typecount[i].count = 0;
